feat: migração das rotas de redefinição para o módulo de administradores
