+++ back/src/administradores/schemas/tokenDeConfirmacao.schema.ts
