+++ back/src/usuarios/usuarios.controller.ts
@@ -2,8 +2,7 @@ import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards, Query, Ba
 import { UsuariosService } from './usuarios.service';
 import { CreateUsuarioDto } from './dto/create-usuario.dto';
 import { UpdateUsuarioDto } from './dto/update-usuario.dto';
-import { VerificaTokenDto } from './dto/verifica-token.dto';
-import { RedefineSenhaDto } from './dto/troca-senha.dto';
+
 //import { AuthGuard } from 'src/auth-guard/auth-guard.guard';
 
 @Controller('usuarios')
@@ -37,24 +36,6 @@ export class UsuariosController {
   @Delete(':id')
   remove(@Param('id') id: string) {
     return this.usuariosService.remove(id);
-  }
-
-  @Post('email-redefinicao')
-  async enviaRedefinicaoDeSenha(@Body('email') email: string) {
-    return this.usuariosService.enviaTokenRedefinirSenha(email);
-  }
-
-  @Post('verifica-token')
-  async verificaToken(@Body() verificaTokenDto: VerificaTokenDto) {
-    const { token, email } = verificaTokenDto;
-    return await this.usuariosService.verificaToken(email, token);
-  }
-
-  @Post('redefine-senha')
-  async trocaSenha(@Body() redefineSenhaDto: RedefineSenhaDto) {
-    const { email, novaSenha, novaSenhaConfirmacao, token } = redefineSenhaDto;
-
-    return await this.usuariosService.redefineSenha(email, novaSenha, novaSenhaConfirmacao, token);
   }
 
 }
